[ARM] msm: timer: Fix 7x30 debug timer register offsets and clock source
The debug timer's register addresses are corrected. To avoid future
confusion, the MSM_GPT_* iomap macros are renamed to MSM_TMR_*.  This
clarifies that they represent the address of the first timer register,
which is not necessarily the first GPT register on all targets.

Additionally, the debug timer frequency is changed to 24.576 MHz, since
the 7x30 uses LPXO as its debug timer clock source.

Change-Id: If080a83505deb7b371ba38f2c7e30de4f12519a8
Cc: Matt Wagantall <mattw@quicinc.com>
Signed-off-by: Dima Zavin <dima@android.com>