USB: core: Clear hnp_supported flag upon device disconnect
hnp_supported flag is set for a USB bus when a HNP supported device
is connected on OTG port.  This flag is used to schedule and cancel
HNP polling work.  Clear this flag upon device disconnect.  Otherwise
HNP polling work is scheduled when another device is connected on OTG
port though the device is not a OTG device.

Change-Id: Ib0cdb165a0a8e4f59464a9cf8a433377172588e1
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>