msm: copper: Add BIMC and NOC register interface
Add the BIMC and NoC register interface to control
bus arbitration and configuration settings for copper.
BIMC (Bus Integrated Memory Controller) and NoC (Network
on Chip) are new types of interconnects for Copper.

QoS registers on both NoC and BIMC are used for
bandwidth arbitration. Apps owns the QoS registers for
Krait masters on BIMC and for MMSS masters on MMSS
NoC.

The generic QoS register interface for BIMC and
NoC has been added in this patch.

Change-Id: I13ce2bf5936255440f0e863b92b2d597127329fb
Signed-off-by: Gagan Mac <gmac@codeaurora.org>