usb: gadget: Add RESUME from XO workaround
This is a workaround to MDM9x15 HW limitation of failure
to wake up from XO shutdown by USB RESUME signal.
To overcome this, a board rework is needed, so the CPU
can wake up directly from D+ line. A GPIO IRQ is used for
that purpose, and the ci13xxx_msm registers for this IRQ.

CRs-Fixed: 359272
Change-Id: Ica148bc4c4eea5b0c1feed396afd617fdb19a9dc
Signed-off-by: Amit Blay <ablay@codeaurora.org>
[sboyd: Drop board-file changes]
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>