mmc: card: adding specific timeout to the sanitize request
since the sanitize request might take a long time, the timer should
be configured to a large timeout.
the timer is per-request timer. it is started when the request is being
issued, and incase the card is not responding or there is another issue
that might cause the SW to halt, the timer expires.
the timer becomes effective when issuing long-duration requests.

Change-Id: Ieca9b27cf08a03cba57e28f6e58151bec3b623dc
Signed-off-by: Yaniv Gardi <ygardi@codeaurora.org>