msm: msm_sdcc: Set default clock gating timeout to 200msecs
With commit c84f15ae, the default clock gating delay in mmc host
init is set to zero. For MSM targets this can cause huge
performance hit as sometimes the clocks are non-local and
requires a request to other subsystem leading to significant
roundtrip delay.

Fix this to have atleast 200msecs delay in turning off the sdcc
clocks after a request is completed.

Change-Id: I5f434cf98373154d817735dba8f5bc8810d27611
Signed-off-by: Sujit Reddy Thumma <sthumma@codeaurora.org>