dt/msmcopper: Add devices for pm8x41 mpps
Add MPP devices for pm8841 and pm8941 PMICs. There are 4 MPPs on
the pm8841 and 8 on the pm8941.

The 'odd' numbered MPPs have no current sink. The 'even' numbered
MPPs have no analog output buffer support.

The qpnp-pin driver automatically detects the
type and limits the capapabilities of these devices, so no
special Device Tree bindings to capture this information
are necessary.

Change-Id: Ia4d4a5f6ecd2a763112153f35c3e173b0a812022
Signed-off-by: Michael Bohan <mbohan@codeaurora.org>