msm: io: Change the static iomappings to be shared
With 3.4 kernel the static iomappings can be shared with the ioremap
mappings. If ioremap is called with an address for which a static
mapping already exists, then that mapping should be used instead
of creating a new one.
However, the MT_DEVICE_NONSHARED flag prevents this. Hence, get rid
of this flag.

Change-Id: I0347f8d319510bc437bbcf20a228fb5004fec74c
Signed-off-by: Rohit Vaswani <rvaswani@codeaurora.org>