msm: board-8064: Disable pull-down in requests for S7 and L27 regulators
Change the configuration of PM8921 regulators SMPS 7 and LDO 27
so that requests sent to the RPM for these regulators have pd=0
(i.e. pull-down disabled).  This modification reduces leakage
current that is seen during XO shutdown when S7 and L27 are
disabled but S3 is kept on.  The outputs of S3 and L27 are wired
together so the S3 output voltage is continuously dropped across
the L27 pull-down resistor if it is enabled.  S7 is the parent
supply of L27 so the same problem is seen if its pull-down
resistor is enabled.

Change-Id: Ic1b380edc9f60218f2012f2a4bb67467dc099005
CRs-Fixed: 369947
Signed-off-by: David Collins <collinsd@codeaurora.org>