msm: kgsl: Improve detection of a type3 CP packet
Currently we just check the most significant two bits of a dword
in the ringbuffer or IB are set to decide if the data is a type3
packet header.  This has proven to be problematic, as it turns out
that the top two bits are often set in nominal data.  Enhance
the type3 packet check to verify that various reserved bits in
the dword are 0 as expected.

Change-Id: Ic0dedbada5eff8353bc64da9eefd9919de127fa4
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>