usb: msm_otg: Fix host mode suspend
msm_otg_set_suspend relies on the state machine to determine if
a state transition will properly put the transceiver in low power
mode. However, if it succeeds, and is called again to perform the
same operation, the resulting pm_runtime_resume will unnecesarily
bring the device out of low power mode before suspending it again.
This scenario can happen if the otg device is already runtime-
suspended, and a system suspend calls this function again, causing
a temporary resume. However, the controller IRQ will be triggered
due to a port change and will cause the system suspend to abort.

The fix is to simply return if the function is being called to
suspend but the transceiver is already suspended.

CRs-fixed: 366134
Change-Id: Ie391b6b67beeff0030e4fc4b94a54f0e84e824bc
Signed-off-by: Jack Pham <jackp@codeaurora.org>