msm: camera: Fix lens lock issue in actuator.
The actuator is not able to move beyond one region boundary.
This is because when there is region boundary between current
lens position and destination lens position, the lens is first moved
to region boundary and then moved to the destination target position.
Once region boundary(target lens position in this case) is reached
we should not assume it is the end of total lens movement. So
condition check (curr_lens_pos == target_lens_pos) is wrong. Hence
removing the inaccurate condition check.

Change-Id: I0346277b4e9a3f1e0e4ce7bd11c8dbaacfc288f7
Signed-off-by: Rajakumar Govindaram <rajakuma@codeaurora.org>