msm: smd_pkt: Register platform driver with the SMD channel name
Currently during the open operation, platform drivers are registered
to detect the arrival of SMD channels in the system. The platform
drivers are registered with "channel_name.edge" extension. The platform
devices registered by the SMD driver contains the edge information in
the id field. But the platform device name does not reflect "name.id"
format. This causes the probe function not to be invoked. Hence register
the platform driver with just the SMD channel name.

Change-Id: Ia957b62336ddcfb8aa80c858e1da40209cecd2d5
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@codeaurora.org>