msm: acpuclock-krait: Clean up error paths in initialization code
BUG_ON() may be appropriate for critical clock initialization that will
prevent proper booting, but it is not appropriate for functions called
in hotplug notifiers while bringing additional CPUs on line. There, it
is better to just fail the hotplug rather than crashing on purpose.

Fix this by cleaning up the initialization error paths so that functions
called in the hotplug path return errors and cleanly rollback state.

Change-Id: I7b842956ea12ebb94c7c5b05de7a6dddfa298617
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>