arm/dt: msm8974: Add labels spmi_bus and rpm_bus
Add labels spmi_bus and rpm_bus to the devices qcom,spmi@fc4c0000
and qcom,rpm-smd respectively. Replace the appropriate addresses in the
related files by the corresponding labels. Remove the address
dependencies in the dtsi files, and therefore, remove the platform
dependency. Change only the address in the label for a given platform.

Change-Id: Ib93af7f6fc5bdf0da2ddb2fa7aabde0de0c87e6e
Signed-off-by: Varad Deshmukh <varadd@codeaurora.org>