msm: board: Set the correct 2D core clock frequency in 7x30
The 2D clock was synchronously tied to the AXI bus in frequency on
the 7x30 platform. So its was never set before. Now, with the
updated clock APIs, this made the clock to be enabled before it is
prepared. This change ensures the clock is prepared before it is
enabled.

Change-Id: If05be2a4e4742687cb5fe7de2f27136be69b2dc3
CRs-fixed: 375424
Signed-off-by: Ranjhith Kalisamy <ranjhith@codeaurora.org>