msm: pil-q6v5-mss: Combine pil-mba and pil-q6v5-mss drivers
The pil-q6v5-mss and pil-mba drivers are two drivers for the
modem subsystem that load the first and second stage elf images
of the modem. First pil-q6v5-mss loads the mba image which then
authenticates the modem image that pil-mba loads. To properly
express the dependency that the mba image be loaded before the
modem image, we make the modem image depend on the mba in PIL.

When we move to subsystem_get()/subsystem_put() we'll only have
one set of powerup/shutdown ops for a subsystem and so having a
dependency between the two images will no longer be expressible
via a depends field. We'll also no longer have a way to reference
a particular PIL image by string name since all interaction with
PIL will be done via pointers instead of strings.

Combine the two modem PIL drivers on 8974 because in the future
design of PIL and SSR we'll no longer have a way to tie together
pieces of a subsystem via a registration based framework like PIL
is today.

Change-Id: I421f733c180dec37132431fd3682dc02c4113af2
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>