msm: Kconfig: Update SUBSYSTEM_RESTART config
Remove the depends on ARCH_* because we shouldn't have to update
this line whenever a new arch is added. Also, expand on the
Kconfig text to describe new functionality that this config
enables.

Unfortunately, this means sysmon is now exposed as default y
everywhere that subsystem restart is enabled, so we must
explicitly enable sysmon for those architectures and remove the
default y from sysmon too.

Change-Id: I4a1cf701cb0a4492639471fb562625a461c7870c
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>