msm: clock-rpm: Store the RPM clock resource keys in the rpm_clk struct
RPM clocks may have key-value pairs with differing keys.
For example, rate-settable RPM clocks have 'KHz' as a key,
and branch clocks have 'Enab' as a key. To accomodate for
this, store the relevant key in the rpm_clk structure
and pass the key in via the clock defining macros.

Change-Id: I0dc051410e161d1c3f8199d9b17e4983a55185f3
Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>