media: dvb: dvb-core: Pass ring-buffers info to kernel-demux
In order to support mapping of the ring-buffers of
demux-device directly to HW output, kernel-demux needs
to get the information of the ring-buffers. In addition,
when these buffers are mapped to HW output, kernel-demux
needs to be notified when data is read from the buffers
so that the respective HW FIFO can be updated accordingly.

Change-Id: I29e2e89476dd343dbebb440c6feb955a76fd0bd1
Signed-off-by: Hamad Kadmany <hkadmany@codeaurora.org>