msm: camera: Map/Unmap the VPE I/O buffers.
Currently the input buffers and the output buffers configured to VPE
are directly picked up from the v4l2 buffer queue, when a particular
frame is sent to VPE for processing. This adds a condition that the
input/output buffer should be present in the v4l2 buffer queue at all
times. If the user wants to do some processing (for eg: Image
stabilization) on the frame and then send it for VPE processing, this
can not be achieved, since the frame would be dequeued by the user for
processing. Add support to map these buffers when they are sent for
processing and unmap them, when the user is receiving the processed
buffer. Add two generic APIs msm_mctl_map_user_frame and
msm_mctl_unmap_user_frame to perform the necessary map/unmap
functionality.

Change-Id: I5f1489b38844b5d2bcacb67a17f6d63e1aced342
Signed-off-by: Kiran Kumar H N <hurlisal@codeaurora.org>