v4l2: Add mutex to streamon() and dqbuf()in v4l2 framework.
When v4l2 streamon() is called the framework moves to streamon state,
after queueing the buffer. A state check is performed during dqbuf().
Sometimes when dqbuf() is called streamon() function has not completed
setting the state transisiton flag. This results in failure to dqbuf() as
the framework has not moved to streamon. This patch adds a mutex
around state transitions and dqbuf() calls.

Change-Id: I0eadc4aaffbed62facd7dd250893e50448d7e52b
Signed-off-by: Ashray Kulkarni <ashrayk@codeaurora.org>