msm: dcvs: get rid of the pending freq array
Only one pending frequency should be set, there is no need to
maintain an array of intermediate requests.

Cleanup the code and remove new_freq[] array and the index to
it. Introduce pending_freq a single entity that carries the
latest frequency request. Use the core's cpu_lock to protect
it.

Change-Id: If2bc47822d70f97f502df857f01b81134ce669ea
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>