msm: dcvs: update dcvs if the governor limits change
We had seen issues where dcvs goes out of sync with the actual
freq the cpu is running at. The root cause was if the userspace
changes the limits on the governor, the governor ends up changing
the frequency without notifying dcvs.

Provide an api for the governor to call dcvs when a frequency change
happens.

Change-Id: Ifb33fac9cdde8535e274b7c91534dc44cc19dbe7
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>