msm: 9625: Initialize RPM SMD driver.
Add an init function in the board file to initialize the RPM SMD driver.
Also modified the rpm smd driver to allow standalone mode for 9625 only.
Standalone mode is generally needed during early development on a new
chip so adding it for 9625, but removing it for 8974 since the chip is
beyond initial bring up stage.

Signed-off-by: Girish Mahadevan <girishm@codeaurora.org>
Change-Id: I8913adaccca941bfc1088d4bb5c4264211596c43