diag: Fix check for reissue of smd control channel read
If there are no messages of type DIAG_CTRL_MSG_REG in the
data being processed on the smd control channel, then the
channel is not rechecked after processing is completed to
see if the channel has more data to be read.  Add check
to ensure the channel is checked for more data regardless
of the types of messages being processed.

CRs-Fixed: 405424
Change-Id: I5580aaa29db973e1ef02b5db81feb4da4b3b09b8
Signed-off-by: Dixon Peterson <dixonp@codeaurora.org>