ASoC: msm8974: Fix ADM memory map callbacks.
Set the token field of the APR packet with the port ID
and the destination field with the COPP ID. With the
token field set to the port ID the wait thread is
correctly signaled by the memory map callback. Add
error recovery if ADM memory map command fails.

Change-Id: If16fac2e4a52a924b15d44ab9e29097cfd9362cf
CRs-fixed: 404941
Signed-off-by: Ben Romberger <bromberg@codeaurora.org>