msm: lpm_resources: Modify L2 resource handling
On targets that don't have L2 SPM, the L2 resource is not valid. Modify
the code not to return that the L2 resource is not beyond limits in such
cases and also prevent setting the L2 SPM mode if the L2 resource is not
valid. This change is needed to allow power collapse to go through on
targets that don't have L2 SPM.

Change-Id: I0abc6c3a796d74633f417ce8ad9624d179fa8b27
Signed-off-by: Girish Mahadevan <girishm@codeaurora.org>