usb: msm_otg: disable IRQ during resume
On targets supporting the secondary low-power mode interrupt,
when it is fired it will trigger the resume sequence. During
this time as the controller is being brought out of suspend,
the normal USB interrupt may also fire before we have finished
with the routine, and in the interrupt handler would overwrite
the variable that recorded the first interrupt. This will lead
to unbalanced enable_irq/disable_irq calls and subsequent
interrupts will be inadvertently ignored.

A simple fix is to disable the USB IRQ for the duration of the
resume sequence, and re-enable it right before exiting.

Change-Id: Ia24a1adbe45b490a9606e75ac551f9286face8dd
Signed-off-by: Jack Pham <jackp@codeaurora.org>