rmnet_smux: Wake netif queue after low-watermark notification
Currently, RMNET will restart the TX queue after receiving a
low-watermark notification from SMUX.  This just clears the XOFF bit,
but does not reschedule the TX thread in the Linux TCP/IP stack.  This
means that the next TX operation will not take place until the thread is
scheduled by some other means which may take up to 5 seconds.

Instead, wake the queue which clears the XOFF bit and schedules the TX
thread to allow transmission to continue immediately.

CRs-Fixed: 412758
Change-Id: I322ba4c564c3b042f5d6bf59514c019719d5b38a
Signed-off-by: Eric Holmberg <eholmber@codeaurora.org>