power: pm8921-charger: fix resume SOC check problem
The current code assumes that the battery state of charge (SOC)
increases or decreases by one percent. This is not so. The check
for resume charging could fail and we might enter a condition when
the SOC is lower and we might not resume charging.

Fix it by checking whether we have crossed resume SOC threshold.

While at it there is also an issue where we might not resume
charging if the charger was plugged out before end of charge but
above vbatdet threshold, if the charger is inserted back again
it might not resume because the current algorithm seeks for a
drop in SOC. Update the check to only lookout for a SOC less
than resume percent and if the charging is not in progress.

Change-Id: I8cac84296ddb1e323f28e2476f78d0190f4a1b73
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>