msm: clock-mdss-8974: Reference-count the PLL enable/disable ops
The DSI PLL enable and disable operations are common to the
pixel and byte clocks. This poses a problem because turning off
the pixel clock first will leave the byte clock RCG on with a
source PLL that is turned off (causing the clock to be 'stuck
on'). A bug in clock-8974 was causing this problem to be hidden
- the byte clock enable and disable ops were being used for both
the pixel and byte clocks. Correcting that bug exposed _this_
problem.

To address this, add reference-counting to the PLL enable
and disable operations.

Change-Id: Idc88abcfede940d3cd0433627ac8dfc8b22549ca
Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>