msm:9625: Move memory map information from the bootloader to kernel.
Update the board file to scan the device tree to get the address and
size of the memory holes after the kernel start address. These memory
holes are removed using memblock-remove.

Any memory not accessible to the kernel below the start address
should be not be given by the bootloader to the kernel.

Signed-off-by: Neeti Desai <neetid@codeaurora.org>
Change-Id: Ic38b88b32095ffde3f4e7db2e1b5c2caea651a8c