mtd: msm_qpic_nand: Remove the unnecessary (and incorrect) print.
The details of the number of partitions found and MTD partitions
created are already printed out to dmesg log by the functions
parse_mtd_partitions and add_mtd_partitions. More importantly
this line always prints the number of partitions found as 0 which
is incorrect. On successful completion, mtd_device_parse_register
returns 0 and doesn't return the number of partitions found.
So remove this print completely. Also remove the variable n_parts
and use err variable which is more appropriate.

Change-Id: I468f79d1d8bcaf80f24dc55c5db0bad93a66f55b
Signed-off-by: Murali Palnati <palnatim@codeaurora.org>