msm: vidc: Use proper msm_vidc_core when handling SYS_ERROR
When handling SYS_ERROR use device_id to determine the core rather than
find the core via the instance (contained with in session_id).  The
firmware does not respond with the proper session_id in cases of
SYS_ERROR.

When handling SESSION_ERROR pass through the session_id returned by the
firmware, so that the instance can be cleaned up properly.

Change-Id: Ic38feb0878579403a1c8e108f18d46bea2891795
CRs-Fixed: 419792
Signed-off-by: Deva Ramasubramanian <dramasub@codeaurora.org>