dvb: dvb-core: Use kthread instead of workqueue to process DVR input
Existing code uses workqueue for DVR input work which constantly waits
for new data written to the input buffer. This causes the kworker
executing DVR work to be fully occupied by DVR device and does not
allow other drivers using workqueues to use that kworker. To solve
this switched to kernel thread which is solely used for DVR device.

Change-Id: I7e501d6e1e6668d9f4d275cf5e2d46cebe3aa68e
Signed-off-by: Hamad Kadmany <hkadmany@codeaurora.org>