arm/dt: msm8974: Remove always-on and parent properties for RPM regulators
Remove regulator-always-on and parent-supply properties for all
MSM8974 qcom,rpm-regulator-smd device nodes.  Also remove the
associated properties qcom,init-current and qcom,system-load from
always-on regulators.

The RPM processor guarantees that always-on system regulators
cannot be disabled.  It also manages the voltage and mode for
these regulators during system active and sleep state in order to
minimize power consumption.  Removing regulator-always-on and the
associated qcom,init-current and qcom,system-load properties for
these always-on regulators ensures that the Apps processor does
not interfere with the RPM processor power saving efforts.

The RPM processor also handles parent-child relationships between
top level SMPS regulators and the LDO regulators that are
sub-regulated from them.  In particular, it ensures that a given
parent regulator is enabled immediately before any of its child
regulators are enabled.  It also ensures that the voltage set
point of the parent regulator is sufficiently high to handle the
dropout voltage for its children.

Removing parent-supply properties causes the Apps processor to
not send redundant parent regulator requests to the RPM
processor.  This leads to a reduction in regulator control
latency.

Change-Id: I0c2bb663471a492e908d5296d1e8acc5f68a782b
Signed-off-by: David Collins <collinsd@codeaurora.org>