msm: clock-8974: Rework the HDMI PLL implementation.
The extpclk_clk_src RCG uses the HDMI PLL as its only source,
which is a rate-settable PLL. However, it is not possible to
call clk_set_rate on this source PLL in the RCG set_rate op,
since that op is invoked in an atomic context. To solve this,
replace the single RCG set_rate op with custom clock ops that
include a new rcg_clk_set_rate function.

Also, remove unnecessary locking in the HDMI set rate and
enable/disable functions.

Change-Id: Ibddb00b5c402ea027f76fe38a99703cf32ed97d3
Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>