power: qpnp-bms: read selected BMS IADC channel
In the current implementation, the BMS assumes that the BMS IADC is
reading from an external rsense. This is not always the case, since
the hardware could be reading from the internal batfet resistance.

Register the IADC1_BMS_IADC peripheral in the BMS driver to read the
selected IADC channel and make the information available to the BMS
driver.

Change-Id: I329914134b0b8c02cf40aca16bb28a66a90e3d5f
Signed-off-by: Xiaozhe Shi <xiaozhes@codeaurora.org>