msm: clock-8974: Use device tree generated names for PIL devices
OF_DEV_AUXDATA is used to rename the names of PIL devices and
the new names are registered in the clock lookup table for
the corresponding clocks.

However, this is not required now for two reasons. First, the
device tree created device name is clean enough. Second, the
OF_DEV_AUXDATA may not be supported in Device Tree in the future.

Therefore, use the device name created by device tree from now
on in this patch.

Change-Id: I45aa9c9760a15627d802da7e400a98a3fe72ab86
Signed-off-by: Tianyi Gou <tgou@codeaurora.org>