arm/dt: 9625: Set swfi_latency for vote to CPU DMA latency
While USB cable is connected, USB driver is waiting for interrupts for
resuming of data/control transfers.
Lpm levels of the msm9625 defines 100us as the max required latency for
the CPU to wake up from idle state and still be able to handle interrupts
correctly. (Please see arch/arm/boot/dts/msm9625-pm.dtsi)

This change adds a vote to the cpu-idle subsystem so will ensure that
the CPU will not enter to specific idle state from which the penalty
wake up time is larger then 100us.

Change-Id: I1bc6cfdf0dc1036886242136eef61f89ae702e5d
Signed-off-by: Ido Shayevitz <idos@codeaurora.org>