8974: i2c-qup: Correct the source clock-frequency
Source clock is incorrectly set to 24MHz on 8974. The driver
calculates the value to set in the QUP-clock-dividers from two
DT entries: bus-frequency and source-clock-frequency. The
correct source clock rate on 8974 is 19.2MHz and not 24MHz,
thus, the divider value is wrong, which results in a wrong
I2C bus frequency.
This patch corrects the source clock frequency in the DT to
19.2MHz.

Change-Id: I4cfc5fda3c0f5e403d8a5fa94d24cdc30f9efc0d
CRs-Fixed: 425744
Signed-off-by: Gilad Avidov <gavidov@codeaurora.org>