camera: Flush event queue for config node before unsubscribe.
A msm_isp_event_ctrl structure may be attached to a v4l2_event
entry. Before unsubscribing from the event queue, we need to
go through the queue and flush all the entries.

If we leave it to v4l2_event_unsubscribe, msm_isp_event_ctrl
will be leaked.

CRs-Fixed: 433808
Change-Id: I4e7187093c6aa11c64d83a00c310f773a6067f14
Signed-off-by: Shuzhen Wang <shuzhenw@codeaurora.org>