defconfig: make cfg80211 statically linked
Prior to the current Android release the WCN wlan drivers and ATH wlan
drivers used different versions of the cfg80211 module, so cfg80211
was made a loadable module, and changes were made to the Android
framework to load the correct one based upon which driver was being
used when WLAN was enabled.  For the current Android release we don't
want to make unnecessary changes to the framework, so only support a
single built-in cfg80211 module.  All wlan drivers on the current
Android release will need to use this module.

Change-Id: Iea4268a83a64f67a70f6ee9c35abbd2ddfcf8a1f
Signed-off-by: Jeff Johnson <jjohnson@codeaurora.org>