msm: board-8064-gpiomux: clean up of AP2MDM GPIO configurations
Fix drive strength of each of the ap2mdm gpios cleanup so as to avoid
any leakage current and incorrect state. The ap2mdm gpios are
using the default MUX configuration to set them as input and we change
the direction of the gpio as required. Having high drive strength for
these gpios will end up with more leakage current.

CRs-Fixed: 409427
Change-Id: I3e607b2e993c8dbe6b34e32a436749a348686380
Signed-off-by: Taniya Das <tdas@codeaurora.org>