ASoC: msm: Fix the AUX PCM ports spurious start and stop
If the shutdown of the dai driver is called before the
prepare for the AUX PCM port then the port is closed down
and updates the incorrect aux_pcm_count.
This results in spurious start and stop of the AUX PCM ports.

To track whether the port is started or not update the
flag status_mask of the dai data.

CRs-Fixed: 416502
Change-Id: I101f61355bd4b82aeea7b0ed6658327c1a528f62
Signed-off-by: Aviral Gupta <aviralg@codeaurora.org>