msm: audio: Don't add memory map failed handle to memory map list
During cases where memory map is failed, before "region"
data structure is freed, we need to delete the region list
entry which is added to the begining of the list. Otherwise
this memory will be accessed in ion_remove when iterating
through the list to delete the nodes, which will lead to
device crash.

Change-Id: I002aff8d7d14dfc714adc55c3f21c25dc9f75797
CRs-fixed: 443453
Signed-off-by: Phani Kumar Uppalapati <phanik@codeaurora.org>