arm/dt: msm8226: Add an S1 corner regulator and consumer-supplies entries
The MSM8226 clock driver needs to vote for the regulators that
control the CX and SR2 PLL power rails. This driver does not
employ the device tree; therefore, add the consumer-supplies
property to the corresponding regulator entries.

The clock driver uses corner voting on the S1 regulator. Add
a stub regulator to allow for this.

Change-Id: Ib011a73377d997a03dc5b8be218ba1696dabe1ad
Signed-off-by: Patrick Daly <pdaly@codeaurora.org>