msm_serial_hs : Add support to configure GPIOs
This change adds support to configure/unconfigure GPIOs connected
to HSUART peripheral if GPIOs related information is provided with
enumerated HSUART port. GPIOs would be configured with active
configuration when HSUART port is opened and would be configured
with suspend configuration when HSUART port is closed by application
provided GPIOMUX settings are available for those GPIOs with platform
specific configuration.

Change-Id: I18dcb46df2bed8f6b1980b789d91abfa4b2898c2
Signed-off-by: Saket Saurabh <ssaurabh@codeaurora.org>