msm: smd_tty: correct wakelock cleanup
The read available wakelock for a port is not correctly cleaned up in
smd_tty_close for the case where the underlying SMD transport open call
(smd_named_open_on_edge) failed.  If the port was reopened after this
failure, then the global wakelock list would be corrupted and could result
in watchdog crashes.

Fix this problem by changing the conditions under which the read available
wakelock is cleaned up in smd_tty_close so that all cases are handled
properly.

CRs-Fixed: 452589
Change-Id: Ia1cef0c94ce1ed36087ef618b8dcc2426d313fdb
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>