arm/dt: Update DT VADC documentation
Document the pre-div configuration support for
div1, div3, div4, div6 and div20. The pre-div configuration
is used by post scaling functions to determine the scaling
to the applied to the input channel before returning the
final physical converted result. The raw ADC code is not
affected.

Add an example on how to configure an MPP as an analog
input to an AMUX. MPP's is not part of the ADC block.
However MPP's are used when any analog input signal
is to be read. MPP block supports configuring an
MPP as an analog input and read from one of the
pre-mux VADC inputs.

Change-Id: I1e5f3ffbc5bfdf5f637fd68fb63e3a23adebd25c
Signed-off-by: Siddartha Mohanadoss <smohanad@codeaurora.org>