msm: clock-9625: Support measurement on both v1 and v2 hardware
IPA and QPIC clocks measurement is only supported on v2. With the current
implementation of supporting these clocks measurement on v1 too, the
clock test will fail on v1.

To fix this problem, constructing the measurement mux during runtime based
on the hardware version so that only measurable clocks are exported.

Change-Id: I27b4cfc8ea59727ec6204dc325b987e7cddf2b18
Signed-off-by: Tianyi Gou <tgou@codeaurora.org>