msm: pil-q6v5-mss: De-assert MBA proxy votes in modem_shutdown/stop()
Although the MBA doesn't have a subsystem shutdown function, it does
have proxy votes that need to be removed as part of modem_shutdown()
and modem_stop() if the scheduled work to do so hasn't expired yet.
Fix this by calling pil_shutdown() for the MBA within these functions.

This prevents proxy votes from remaining asserted longer than
necessary if the modem is booted up and shut down back-to-back.

Change-Id: Ic52a22cfac6479fe85cd1f5f8d6773d4df724900
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>