diag: dci: Process DCI packet during USB disconnect
Diag driver drops the packet from the user space when the logging
mode is USB and the USB is disconnected. As a result the DCI
packet coming from the user space is dropped when USB is disconnected.
This change adds a condition to allow DCI data to be processed in this
condition. Also, reset the smd write retry count for every log
equipment ID.

Change-Id: I814d06bc0d2fd755d3c7223ff415b6f38dea8613
Signed-off-by: Ravi Aravamudhan <aravamud@codeaurora.org>