msm: kgsl: Copy data back to user in case of error
In case of an error, copy back any data that KGSL might
have changed on purpose.  This is to prevent the user from
ignoring the error and using the values that were passed in.
This means that in error cases, KGSL needs to be mindful
and reset certain values (such as a buffer size to 0) before
returning the error to the user

Change-Id: I9201ce4361c9fd1d6477f18c67f5c651e8408bd2
Signed-off-by: Carter Cooper <ccooper@codeaurora.org>