power: qpnp-charger: fix download mode on CDP
The qpnp-charger driver enables the USB
suspend bit and lowers the input current
limit of USB during the restart sequence.

On some platforms which do not have a
battery present this may cause a UVLO
reset and thus cause IMEM to be reset.

Fix this not setting the USB suspend bit
and lowering the input current limit if
use_default_batt_values is set and no
battery is present.

Change-Id: I5c70e1aa137c78d5c43feaa9602b7b27fbd1eb47
Signed-off-by: David Keitel <dkeitel@codeaurora.org>