usb: dwc3-msm: Fix spurious ID interrupts during MHL detect
When used with the MHL SiI8334 chip, ID low event kicks off the
MHL discovery mechanism. However, during discovery, the MHL chip
disconnects its USB ID output line to the PMIC and leaves it
floating, which the PMIC will sense as a low->high transition.
If the plugged-in cable turns out to not be MHL, the USB ID line
is reconnected and PMIC again will sense a high->low transition.
This falsely re-triggers MHL discovery to happen again endlessly.

Fix this by disabling the PMIC ID interrupt for the duration of
the discovery phase. Also ensure that the IRQ handler only handles
the case when ID line actually changes.

CRs-fixed: 466887
Change-Id: I7cbd98be93de60338a930e9e1a11e7678f0ae020
Signed-off-by: Jack Pham <jackp@codeaurora.org>