power: qpnp-bms: fix invalid shutdown soc
The data registers that are used to store the shutdown SoCs in
A family chips had a reset state of 0x0. This changed in B family chips
to be 0xFF. 0xFF is the value the BMS driver uses to store 0 to
distinguish it from the reset state. Thus, the new reset state of 0xFF
is confusing the BMS driver to always force a shutdown SoC of 0%.

Fix this by removing the 0xFF check in read_shutdown_soc.

Change-Id: Ic0e4720514324468be7f5871c18dfc754a87feb3
Signed-off-by: Xiaozhe Shi <xiaozhes@codeaurora.org>