Squashfs: fix corruption checks in squashfs_lookup()
The dir_count and size fields when read from disk are sanity
checked for correctness.  However, the sanity checks only check the
values are not greater than expected.  As dir_count and size were
incorrectly defined as signed ints, this can lead to corrupted values
appearing as negative which are not trapped.

Change-Id: Ife6e238b642825f8d03e94de7bec978e5e57706e
Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>