Squashfs: add corruption check in get_dir_index_using_offset()
We read the size (of the name) field from disk.  This value should
be sanity checked for correctness to avoid blindly reading
huge amounts of unnecessary data from disk on corruption.

Note, here we're not actually reading the name into a buffer, but
skipping it, and so corruption doesn't cause buffer overflow, merely
lots of unnecessary amounts of data to be read.

Change-Id: Ice130b762aef503d0e08f6e8ba209398dea9b2e9
Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>