power: max17050: check version of custom parameters
To update custom parameters without fuel gauge's
POR(Power on reset), define and store the version
of custom parameters to the unused register, 0x20
register. MAXIM confirmed the 0x20 register can be
used for this purpose. By checking this version,
driver can detect the change of the parameters and
update them without POR.

For use of userspace, add sysfs entries which pass the
status of init done and power on reset.

Change-Id: I8c1dafaa90863b1b336696e9a72a44f66f03f7a4
Signed-off-by: choongryeol.lee <choongryeol.lee@lge.com>