UPSTREAM: HID: sony: Simplify LED initialization and eliminate redundant copies when updating LED states
Directly set the initial LED states in the device state struct instead
of copying them from a temporary array.

This allows for the removal of a redundant "x = x" copy loop in
sony_set_leds() that was taking place any time an LED was updated.

It also allows for the simplifying of the parameters in functions dealing
with LED initialization and updates since only a pointer to the sony_sc
struct is needed now.

Signed-off-by: Frank Praznik <frank.praznik@oh.rr.com>
Acked-by: Pavel Machek <pavel@ucw.cz>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
(cherry picked from commit 221399b36535752351d4631606b427daca3dc35c)

Bug: 111431828
Signed-off-by: Kim Low <kim-huei.low@sony.com>
Change-Id: I72d2f6e265f4f98286f676dc4a74e8ed42335694
Signed-off-by: Siarhei Vishniakou <svv@google.com>