fuse: Add support for shortcircuited read/write for files
Add support for shortcircuited read/write for files
when enabled through a userspace init option of
FUSE_SHORTCIRCUIT.

When FUSE_SHORTCIRCUIT is enabled all the reads and writes
to the fuse mount point go directly to the native filesystem
rather than through the fuse daemon. All requsts that aren't
read/write still go thought the userspace code.

This allows for significantly better performance on read and writes
and the difference between fuse and the native lower filesystem is
negligible.

Bug: 30222859
Change-Id: I49e21b77813595c2faec6fcba38a74e8f686d020
Signed-off-by: Nikhilesh Reddy <reddyn@codeaurora.org>