Fix generic cgroup subsystem permission checks
In 53b5e2f generic cgroup subsystem permission checks have been added.
When this is been done within procs_write an empty taskset is added to
the tasks css set. When a task later on migrates to a new group we see
a dmesg warning cause the mg_node isn't empty (cgroup.c:2086). Cause
this happens all the time this spams dmesg.

I am not really familiar with this code, but it looks to me like
adding the taskset is just a temporary action in this context.
Therefore this taskset should be removed after the actual check. This
is what this fix does.

This problem was seen and the fix tested on x86 using l-mr1 and master

Change-Id: Id35fc690d3c3be8f064915ce99aee8f0a68148a3
Signed-off-by: Christian Poetzsch <christian.potzsch@imgtec.com>
Signed-off-by: Andres Oportus <andresoportus@google.com>