selinux: removed redundant headers from objsec.h
flask.h was included in the above file. This is autogenerated header
which gets produced during compilation of selinux. This causes issues
for other module (PFK) which includes objsec.h and compiles after
selinux. Since removal of the above headers does not break compilation,
they seem to be redundant and can be removed.

Bug: 68721442
Bug: 68995224
Change-Id: Ia08bcb83570d7c38c40a46bb4a5ec45e9d81e7d7
Signed-off-by: Gilad Broner <gbroner@codeaurora.org>