ARM: dts: msm: Add partial goods support for msm8909
On certains devices, some of the CPUs within the apps subsystem are not
functional and might be removed from the CPU possible map. To handle this
dynamic list of CPUs, sleep driver shouldn't make any assumptions with
regards to the logical mapping of the CPU.

Remove the assumption on how physical CPUs are mapped onto logical CPUs and
create a reference of sleep status driver in the corresponding
CPU phandles. When the sleep status driver is probed, it would probe
iterating the list of possible CPUs.

Change-Id: I027621e4c0885882f85409b5692a3aef508cf9ce
Signed-off-by: Anil Kumar Mamidala <amami@codeaurora.org>
Reviewed-on: http://mcrd1-22-pc.corpnet.asus/code-review/master/224394
Reviewed-by: allen1_huang <allen1_huang@asus.com>
Tested-by: allen1_huang <allen1_huang@asus.com>