wlan: Enable logp ini by default
Currently WdThread is created and gpVosWatchdogContext is
initialized when logp ini is set to 1.
Now when timer callback is executed, it checks for which thread
this timer belongs to.
Crash is observed while checking the threadID
for WDThread as logp ini is not set and gpVosWatchdogContext is
uninitialized.

Enable logp ini to 1 by default to mitigate the issue

Change-Id: I5bee1de46aab4773c8e91a0c76976b712b8f9707
CRs-Fixed: 998016
Signed-off-by: Devin Kim <dojip.kim@lge.com>