ANDROID: nanohub: fix suspend interaction with IRQ2
platforms with IRQ2 available are not able to suspend,
if non-wakeup events are flowing quite often.

To fix it, we acquire lock on wakeup signal, before
attempting to check for IO activity.

We no longer need to check the activity of worker thread,
because if we have the lock, thread has nothing to do but wait
for us, however it may not yet reach the IDLE state, and
we don't want to lose time waiting for this to happen.

Signed-off-by: Alexey Polyudov <apolyudov@google.com>
Change-Id: Id9c661ec022b8a6b1b53ebdcd7392006d74e40d5
Bug: 28271752
Signed-off-by: Andrew Rossignol <aarossig@google.com>