input: misc: rsb: Switch to devres for probe
Devres simplifies the resource management for RSB, and eliminates some
exit paths during probe failure. So, we switch to that for the initial
allocation of resources.

Also, we add a few #defines for certain commonly used registers.

Additionally, specifying interrupts in the device tree enables
the "irq" field of struct spi_device to get automatically populated
before probe(), so we can directly request_threaded_irq(), instead of
manually requesting the GPIO, and setting it as an IRQ. As a result, we
can remove that specification from the DT as well.

Change-Id: Id56d42e500e45fd00b40230bd5116d7d451228ea
Signed-off-by: Prashant Malani <pmalani@google.com>