input: touchscreen: send WAKEUP event properly in ITE Tech driver
Send WAKEUP event properly in ITE Tech touch driver when device
is in suspend mode.
The driver maintins a boolean state for sending the WAKEUP event.
When user taps the screen for touch-to-wake feature, the irq
handler for ITE Tech touch driver is scheduled after the ITE tech
resume function. Due to this the boolean state gets reset.
So this change handles this use-case properly.

CRs-Fixed: 1020680
Change-Id: Ifc1d00a42ecffe7aad65d99dadcdf170c56129a7
Signed-off-by: Shantanu Jain <shjain@codeaurora.org>