power: charger: fix a bug about system wakeup
When AP is in sleep state and connect usb cable, AP can not wake up. The reason is that the
usb interupt can not be called and can not run wake_lock. The interupt triger should be voltage
falling. However, it is configured to low level. So resolve this error.

Change-Id: I1a7152aed7acb97968504a3737154421011cd3ca
Signed-off-by: l00228880 <l00228880@notesmail.huawei.com>