servicefs: Implement support for async "impulses".
Impulses are short async messages from client to service. Impulses are
received through the normal msg_recv API, however the payload is
transferred upfront and returned with the message description instead of
requiring a separate data transfer phase. In order to keep impulses fast
and to prevent overuse, the payload is limited to four machine words.

Bug: 28279882
Test: As per initial implementation.
Change-Id: Ia672c50d8f3c541906d449801bc0878e4b06aea8