input: misc: rsb: Reduce frequency of events
The RSB appears to be generating too many input events, and sending all
of these up to userspace has an affect of overwhelming the framework.

So, this patch  adds a sleep between successive reads of the motion
registers inside of the threaded interrupt handler. Once the handler
wakes from its sleep, it checks the motion register again. If there is
no motion, the handler exits, and if there is, the new values are read
from the X and Y registers.

The current delay is kept at 9 ms.

Change-Id: Ie5c3a8768daca4e4d7a3e6eeb12665c20f03fa9e
Signed-off-by: Prashant Malani <pmalani@google.com>