msm: mdss: check for split display enable in multiple displays
commit 3bf808d38b78b55eb2bf686555a0c2f371eccdf2 ("msm:
fb: add mdp foot switch for multiple splash displays"
considers DSI0 and DSI1 as separate displays even if
split display is enabled. So the number of displays
are calculated as 2 whereas the driver considers this
as a single display. This causes MDP clocks to remain
ON. This change corrects number of displays calculated
by checking for split display.

Bug: 29387552
Change-Id: I444a9e9e34eb7f6fe1b9d93b7f15aaf81986b607
Signed-off-by: Krishna Srinivas <krisrini@codeaurora.org>
Signed-off-by: Naveen Ramaraj <nramaraj@codeaurora.org>