power: idtp9017: Clean up the work to read the information
We need reading the information only when wireless charging is enabled.
So the work to read the inforamtion need to be scheduled when wireless
charging is enabled. And some functions return its value or error code.
In case of that, it's difficult to check error code with the returned
value. Becase correct value would be negative also. So change to get
the value with pointer argument.

And fix the incorrect calculation of the operation frequency

Change-Id: Ie0231d729b2309b9c4c074cbbdd3662ccd8aa8a7
Signed-off-by: Devin Kim <dojip.kim@lge.com>