input: misc: rsb: Clean up regulator init code
Having the regulator init code execute in a work queue can lead to
certain errors (for instance, the probe function may go ahead an
complete, but the regulator init code will fail; this will lead to a
registered driver and input device, but no real way of talking to the
peripheral, or to recover from this error).

This patch moves the regulator init code to execute synchronously from
within probe() itself.

It also gets rid of a couple of extraneous regulator enable/disable
calls for specific regulators. It instead sends a pointer to the
required regulator through to generic regulator_enable/disable calls.

Change-Id: I9d97e7d6df8f102097656417eb01b27c342a8930
Signed-off-by: Prashant Malani <pmalani@google.com>