qcacld-2.0: Do not make a PMF connection, if supplicant sets it as optional
In earlier android builds PMF was not enabled by default in the supplicant
unlike the latest builds. So even if PMF is set as optional, driver was
mandating it. So if we try to roam to an AP which is not PMF capabale
driver would simply drop the connection.

To address this, allow non PMF connection if PMF is set as optional by the
supplicant and the AP is not capable.

CRs-Fixed: 1018482
Bug: 29605770
Change-Id: Ideecb861dcc362f94a73b47d1a4766d621a38214
Signed-off-by: Srinivas Girigowda <sgirigow@codeaurora.org>