qcacld-2.0: Check consistency in throughput before setting delack
Presently, throughput is measured in every 100ms. Delack is set to 20
if throughput for this time duration is greater then threshold (57 Mbps).
In the case of packet drop, tcp takes more time to rampup.
In this change, check the throughput consistency for three second
and then set tcp delack to 20.

CRs-Fixed: 1028085
Change-Id: I155bc141087104ad1443d250711423d5fa936254
Bug: 29490442
Signed-off-by: Srinivas Girigowda <sgirigow@codeaurora.org>