qcacld-2.0: If set roaming fails return non-positive value
If set roaming disabled is called when roaming is already in
progress the sme_config_fast_roaming api returns failure status
and this status, which is positive value, is returned to upper
layer. This results in wpa_supplicant to wait indefinitely.

To avoid this if sme_config_fast_roaming returns falure,
return -EINVAL.

Change-Id: I1de5276f9fc72b04725f963e44e6168bc119f559
CRs-Fixed: 1029834
Bug: 29941999
Signed-off-by: Srinivas Girigowda <sgirigow@codeaurora.org>