qcacld-2.0: Correctly handle the error conditions in the driver startup
In the hdd_wlan_startup function correct the error conditions in case of
any initialization failure, there is crash instance where in the nl socket is
initialized but not released by the driver incase of vos_open failure.

Release the nl socket during the load failure and also correct the other
error conditions.

Change-Id: I5eb63b42de0f2327b391427dd5b43235103ed06b
CRs-Fixed: 1033220
Bug: 29941999
Signed-off-by: Srinivas Girigowda <sgirigow@codeaurora.org>