mdss: dsi: Enhancement of the TE based ESD approach
The current implementation enables TE based IRQ in resume and
disables only in suspend. This results in many IRQs to be fired
even when there are no active frame updates causing unnecessary
CPU wakeups. To overcome this, following approach has been adopted -
* Enable TE IRQ only when ESD thread is scheduled
* Wait for IRQ to be asserted for a maximum duration of 3 vsyncs.
* If not asserted in the given time period, report panel dead due
  to ESD attack.

In this way, the TE IRQ rate is bound by the ESD interval without
altering the main objective of this approach.

Change-Id: I59fe2ae79e88e5687a45209ebfb1e6b42af110bf
Signed-off-by: Sachin Bhayare <sachin.bhayare@codeaurora.org>