msm: mdss: io_util: add new API to configure regulator opt mode
For certain usecases, it may be possible to lower the current load
requested for a particular regulator. For example, during static
display if the DSI lanes are in ULPS mode, it is possible that the
load on the panel regulators can be lowered. To support such usecases,
add a new API to the io_util library to configure the optimum mode
for a power module.

Change-Id: Ie31ef9a67d321fd89767ba2c7f92d6f3b4ae7e7e
Signed-off-by: Sachin Bhayare <sachin.bhayare@codeaurora.org>