usb: phy-msm-usb: fix usb online prop setting failure
When a usb removal is detected, the vbus power is set to
0mA and the online property is set to 0. If the requested
power is same as a cur_power, the request is simply ignored.
However, since the cur_power is initially 0, the request
for 0mA could be ignored at booting. If the request for 0mA
is ignored, the online property setting is also ignored.
Fix this by setting cur_power to -EINVAL initially.

Change-Id: I69a625b9a52603650a28dce2001b11afca8126ff
Signed-off-by: Devin Kim <dojip.kim@lge.com>