common: sched: Fix information leak in sys_sched_getattr()
We're copying the on-stack structure to userspace, but forgot to give
the right number of bytes to copy. This allows the calling process to
obtain up to PAGE_SIZE bytes from the stack (and possibly adjacent kernel memory).

This fix copies only as much as we actually have on the stack
(attr->size defaults to the size of the struct) and leaves the rest of
the userspace-provided buffer untouched.

Found using kmemcheck + trinity.

Fixes: d50dde5a10f30 ("sched: Add new scheduler syscalls to support an extended scheduling parameters ABI")

CVE: CVE-2014-9903
Bug: 28731691

Change-Id: I2492035ce6c0b41088d5f2be239ed348e9197c12
Reviewed-on: http://mcrd1-22-pc.corpnet.asus/code-review/master/243368
Reviewed-by: Yu-Hao Lin <yu-hao_lin@asus.com>
Tested-by: Yu-Hao Lin <yu-hao_lin@asus.com>
Reviewed-by: Eric1 Lin <Eric1_Lin@asus.com>
Reviewed-on: http://mcrd1-22-pc.corpnet.asus/code-review/master/247085
Reviewed-by: jay_chuang <jay_chuang@asus.com>