swift: USB: security patch for CVE-2015-8816
Subject: USB: fix invalid memory access in hub_activate()

Commit 8520f38099cc ("USB: change hub initialization sleeps to
delayed_work") changed the hub_activate() routine to make part of it
run in a workqueue.  However, the commit failed to take a reference to
the usb_hub structure or to lock the hub interface while doing so.  As
a result, if a hub is plugged in and quickly unplugged before the work
routine can run, the routine will try to access memory that has been
deallocated.  Or, if the hub is unplugged while the routine is
running, the memory may be deallocated while it is in active use.

This patch fixes the problem by taking a reference to the usb_hub at
the start of hub_activate() and releasing it at the end (when the work
is finished), and by locking the hub interface while the work routine
is running.  It also adds a check at the start of the routine to see
if the hub has already been disconnected, in which nothing should be
done.

Change-Id: Iedd627107ba22ede4f1ba35c75676c6d419c6b35
Reviewed-on: http://mcrd1-22-pc.corpnet.asus/code-review/master/244999
Reviewed-by: Eric5 Ou <eric5_ou@asus.com>
Tested-by: Eric5 Ou <eric5_ou@asus.com>
Reviewed-by: JimmyLin <jimmy1_lin@asus.com>