Common: net: Security patch for CVE-2015-2686 ANDROID-28759139
Subject: net: validate the range we feed to iov_iter_init() in sys_sendto/sys_recvfrom
The net/socket.c file, in the Linux kernel before 3.19.3, does not validate certain range data for sendto and recvfrom system calls.
This could allow local users to gain privileges by leveraging a subsystem that uses the copy_from_iter function in the iov_iter interface, as demonstrated by the Bluetooth subsystem.
The fix is to designed to add additional bounds checks for validation.

Change-Id: I7965b4bc4da36a03cc6f819d0630ba611384b150
Reviewed-on: http://mcrd1-22-pc.corpnet.asus/code-review/master/249358
Reviewed-by: joe_tsai <joe_tsai@asus.com>
Tested-by: joe_tsai <joe_tsai@asus.com>
Reviewed-on: http://mcrd1-22-pc.corpnet.asus/code-review/master/249361
Reviewed-on: http://mcrd1-22-pc.corpnet.asus/code-review/master/254756