anthias: security: CVE-2016-3843 (ANDROID-29119870)
Elevation of privilege vulnerability in kernel performance subsystem
CVE	Reference	Severity	Updated AOSP versions
	Date reported
CVE-2016-3843	ANDROID-29119870	High	6.0, 6.1
	Google internal

Elevation of privilege vulnerabilities in the kernel performance
subsystem could enable a local malicious application to execute
arbitrary code within the context of the kernel. This issue is rated as
High because of the kernel attack surface available for attackers to
exploit.
Note: This is a platform level update designed to mitigate a class of vulnerabilities such as CVE-2016-3843 (ANDROID-28086229).
Kernel patches for versions 3.4, 3.10 and 3.18 are required in addition
to the AOSP patches. A code snippet for the kernel patch is provided in
the bulletin patches zip file on Google Drive.
Additional technical details:
Bug	Details
ANDROID-29119870	The kernel performance subsystem is intended for
developers and should not be exposed by default on production builds.
Developers may still optionally expose perf using adb.

The fix is designed to restrict unprivileged use of performance events.

Change-Id: I0a6dc6c8d0b25eba40eac565ccd751009ed8e341
Reviewed-on: http://mcrd1-22-pc.corpnet.asus/code-review/master/248102
Reviewed-by: allen1_huang <allen1_huang@asus.com>
Tested-by: allen1_huang <allen1_huang@asus.com>
Reviewed-by: Carol_Jiang <carol_jiang@asus.com>
Tested-by: Carol_Jiang <carol_jiang@asus.com>