ALSA: seq: Fix race at timer setup and close(CVE-2016-2544)
ALSA sequencer code has an open race between the timer setup ioctl and
the close of the client.  This was triggered by syzkaller fuzzer, and
a use-after-free was caught there as a result.

This patch papers over it by adding a proper queue->timer_mutex lock
around the timer-related calls in the relevant code path.

Change-Id: Idc141bddd97a9429b8d303879613755ac8ec646e
Signed-off-by: soon9.lee <soon9.lee@samsung.com>