ALSA: timer: Fix race among timer ioctls(CVE-2016-2546)
ALSA timer ioctls have an open race and this may lead to a
use-after-free of timer instance object.  A simplistic fix is to make
each ioctl exclusive.  We have already tread_sem for controlling the
tread, and extend this as a global mutex to be applied to each ioctl.

The downside is, of course, the worse concurrency.  But these ioctls
aren't to be parallel accessible, in anyway, so it should be fine to
serialize there.

Change-Id: Id04ea96de41ca2eac65a0e0071f65da59d77bbc6
Signed-off-by: soon9.lee <soon9.lee@samsung.com>