Swift: net: Security patch for CVE-2015-2686 ANDROID-28759139
Subject: net: validate the range we feed to iov_iter_init() in sys_sendto/sys_recvfrom
The net/socket.c file, in the Linux kernel before 3.19.3, does not validate certain range data for sendto and recvfrom system calls.
This could allow local users to gain privileges by leveraging a subsystem that uses the copy_from_iter function in the iov_iter interface, as demonstrated by the Bluetooth subsystem.
The fix is to designed to add additional bounds checks for validation.

Change-Id: Ibb26583a8288534166b6f2a417af7b4c885f7ea0
Reviewed-on: http://mcrd1-22-pc.corpnet.asus/code-review/master/249354
Reviewed-by: joe_tsai <joe_tsai@asus.com>
Tested-by: joe_tsai <joe_tsai@asus.com>