input: misc: paj9124u1: Add sysfs enable support
Add support to control enablement/disablement from sysfs.
If the enable flag is set, we still disable the device during kernel
suspend, but re-enable it when we resume.

If the enable flag is not set, then in addition to disabling the h/w, we
will also not re-enable it on kernel resume.

Change-Id: Iaf6d990905a36191a97c041ab441828b90160faa
Signed-off-by: Prashant Malani <pmalani@google.com>