input: misc: paj9124u1: Don't toggle IRQ in sysfs
The userspace seems to call the sysfs "enable" node once after booting.
This leads to an unbalanced IRQ enable count.

So, this patch simply removes the enable/disabling of IRQs so that it
only happens in suspend/resume. We are quiescing the H/W so IRQs won't
get triggered anyway when the H/W is disabled.

This simplifies the logic and prevents us having to depend on userspace
doing the right thing WRT sysfs calls.

Change-Id: I7dec4399945c275b56afd5e35eafc56a10093556
Signed-off-by: Prashant Malani <pmalani@google.com>