anthias: security: CVE-2016-3892
CVE-2016-3892    A-28760543 QC-CR#1024197    Moderate    3.4, 3.10
May 13, 2016

An information disclosure vulnerability in the Qualcomm SPMI driver
could enable a local malicious application to access data outside of its
permission levels. This issue is rated as Moderate because it first
requires compromising a privileged process.

Additional technical details:

A-28760543 (Qualcomm ref#: CR#1024197)

The memory address of a device object is used in the device!|s name
leading to a possible information leak.
The fix is designed to to use a unique ID for each device instead of
using its memory address to prevent the information leak.

Change-Id: I32b70fcc8c376c98a0ed944b58c361c55c3ea2c0
Reviewed-on: http://mcrd1-22-pc.corpnet.asus/code-review/master/250771
Reviewed-by: allen1_huang <allen1_huang@asus.com>
Tested-by: allen1_huang <allen1_huang@asus.com>
Reviewed-by: Carol_Jiang <carol_jiang@asus.com>
Tested-by: Carol_Jiang <carol_jiang@asus.com>