anthias: security: 2016-08 SP CVE-2016-3894
CVE-2016-3894    A-29618014 QC-CR#1042033    Moderate    3.10    Jun 23,
2016
An information disclosure vulnerability in the Qualcomm DMA component
could enable a local malicious application to access data outside of its
permission levels. This issue is rated as Moderate because it first
requires compromising a privileged process.

Additional technical details:

A-29618014 (Qualcomm ref#: CR#1042033 )

The buffers[i] variable in the buffer_req function is allocated memory
but without initialization, possibly leading to an information leak.
The fix is designed to disable the use of the vulnerable file to prevent
the potential information leak.

Change-Id: Iaac12752202e08260496f87f1b05fbf66b126bb3
Reviewed-on: http://mcrd1-22-pc.corpnet.asus/code-review/master/250809
Reviewed-by: allen1_huang <allen1_huang@asus.com>
Tested-by: allen1_huang <allen1_huang@asus.com>
Reviewed-by: Carol_Jiang <carol_jiang@asus.com>
Tested-by: Carol_Jiang <carol_jiang@asus.com>