[2016-08-partner][CVE-2015-5364] anthias: security: patch for A-29507402
udp: fix behavior of wrong checksums
We have two problems in UDP stack related to bogus checksums :

1) We return -EAGAIN to application even if receive queue is not empty.
   This breaks applications using edge trigger epoll()

2) Under UDP flood, we can loop forever without yielding to other
   processes, potentially hanging the host, especially on non SMP.

This patch is an attempt to make things better.

We might in the future add extra support for rt applications
wanting to better control time spent doing a recv() in a hostile
environment. For example we could validate checksums before queuing
packets in socket receive queue.

Change-Id: I5e81733208b98d59afd8e324e464fc5799b6e343
Change-Id: Id6b8291217c77325fd3817f203695fa3064f5a39
Reviewed-on: http://mcrd1-22-pc.corpnet.asus/code-review/master/250566
Reviewed-by: Carol_Jiang <carol_jiang@asus.com>
Tested-by: Carol_Jiang <carol_jiang@asus.com>