ASoC: msm-cpe-lsm: Initialize the period_wait during driver open
The period_wait is used to block the read from the driver in case data
is not available from hardware. It is observed during stability tests
that the task_list for the period_wait is un-initialized. Adding
initialization for the wait queue at driver open to make sure the
period_wait is always initialized as long as driver is opened.

Bug: 28464507
Change-Id: I0f6cc59464f86424d12ceb38133829d7e5fdb679
Signed-off-by: Bhalchandra Gajare <gajare@codeaurora.org>
Signed-off-by: vivek mehta <mvivek@codeaurora.org>