common: ASN.1: Fix non-match detection failure on data overrun
If the ASN.1 decoder is asked to parse a sequence of objects, non-optional
matches get skipped if there's no more data to be had rather than a
data-overrun error being reported.

This is due to the code segment that decides whether to skip optional
matches (ie. matches that could get ignored because an element is marked
OPTIONAL in the grammar) due to a lack of data also skips non-optional
elements if the data pointer has reached the end of the buffer.

CVE: CVE-2016-2053
Bug: 28751627

Change-Id: I448a418f728b68ed9cae833b27f8157340bb0416
Reviewed-on: http://mcrd1-22-pc.corpnet.asus/code-review/master/253605
Reviewed-by: Yu-Hao Lin <yu-hao_lin@asus.com>
Tested-by: Yu-Hao Lin <yu-hao_lin@asus.com>
Reviewed-by: jay_chuang <jay_chuang@asus.com>