nfc: pn458: Remove wake locks
Userspace takes a wakelock when routing modifications happen. This is
more precise in *when* it is taken and relinquished, and also not as
long as the NFCWAKE kernel wakelock, which should be rendered
superfluous.

This patch removes the kernel wake lock entirely, since it seems to be
eating into battery life.

Change-Id: I2114dfe2f954e1a700f1962e01790412dbd2b70e
Signed-off-by: Prashant Malani <pmalani@google.com>