arm/dt: afsf: Reserved the memory for the framebuffer of crash handler
The crash handler uses some memory for the frame buffer to display
information on screen. So it overrites data over the area. The region
is reseved to avoid the overwriting when the crash handler runs.

Device    | FB address |  size
-----------------------------------------
swordfish | 0xa0100000 | 0x100000 (1MB)
angelfish | 0x8ab00000 | 0x100000 (1MB)

Note that removed-dma-pool needs to be aligened with 1MB

Bug: 31664053
Change-Id: I1d7925e65fed7ee39b9f4b2aae34663ed4c11ba7
Signed-off-by: yongtae3.kim <yongtae3.kim@lge.com>
Signed-off-by: Devin Kim <dojip.kim@lge.com>