input: atmel_u144: Remove wake lock and memcpy
The ts_data buffer is memcpy-ed from curr_data to prev_data.
Since the buffers are persistent, it is probably faster to just swap the
two buffers. The curr_data is going to be overwritten the next time a
touch event happens anyway.

This patch also removes the pm_touch_wake_lock, which shouldn't be
necessary. If the handler has fired, the kernel suspend routines should
ensure that all pending interrupt handlers are executed before
continuing with suspend, so there is no danger of the current handler
not being executed.

Change-Id: I688444b517cf14eac5aed68c7c2feae2c92b3d5b
Signed-off-by: Prashant Malani <pmalani@google.com>