drivers: bcm15602: initial support
This is the Easel PMIC. Uses the regulator framework for each of the
regulators. The PMIC also supports 2 GPIOs which are added to the GPIO
framework. Uses a flag ("PREPRODUCTION") to disable I2C transactions and
uses a software table for register accesses. The flag will be removed
once we have silicon.

Bug: 32340281
Change-Id: Ib17df811cdcf169535e5e417ecc1e7a3fff953d5
Signed-off-by: Trevor Bunker <trevorbunker@google.com>