Expose the wait-for-pingpong state to userspace.
This state indicates whether a frame is pending on the display, and
whether posting a new frame will block. Checking this state is
necessary to avoid somewhat rare, but catastrophic race conditions
due to kernel thread scheduling issues causing the display driver
to miss the vsync deadline. Without this check there is no fool-
proof way to determine whether it is safe to post a frame to the
display driver without blocking until the next vsync.

Change-Id: Iba9745a061de046a9d34ce741a18546616ea2ef5
Reviewed-on: https://dreamos-internal-review.git.corp.google.com/5978
Reviewed-by: John Bates <jbates@google.com>
Reviewed-by: Anatol Pomazau <anatol@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>