input: atmel_u144: Clean up needless global var
This code removes an unnecessary global variable that is being used to
reference mxt_data. This is either already available through a function
parameter, or that parameter can be added. There is no need to maintain
a separate global variable for this.

While this doesn't lead to any performance improvement, it's just better
to not have that many needless global variables.

Change-Id: I5be662f789b7c710fd1250642e25de573c976fd3
Signed-off-by: Prashant Malani <pmalani@google.com>