wlan: Increment scan failure counter only on connection in progress
Presenly, along with connection in progress scenario, scan failure
counter is also incremented on BTcoex operation is in progress,
leading counter to reach more than 35 not due to connection in
progress state but due to BTCoex mode in progress.

Hence increment scan failure counter only when connection is in
progress.

Change-Id: I1f770125dc8ee5297b5ee022b4f4f1481a621b25
CRs-Fixed: 1066610