[2016-09-partner][CVE-2016-0758]KEYS: Fix ASN.1 indefinite length object parsing
[2016-09-partner][CVE-2016-0758]KEYS: Fix ASN.1 indefinite length object
parsing

This fixes CVE-2016-0758. In the ASN.1 decoder, when the length field of
an ASN.1 value is extracted, it isn't validated against the remaining
amount of data before being added to the cursor. With a sufficiently
large size indicated, the check: datalen - dp < 2 may then fail due to
integer overflow. Fix this by checking the length indicated against the
amount of remaining data in both places a definite length is determined.
Whilst we're at it, make the following changes:

(1) Check the maximum size of extended length does not exceed the
capacity of the variable it's being stored in (len) rather than the type
that variable is assumed to be (size_t).

(2) Compare the EOC tag to the symbolic constant ASN1_EOC rather than
the integer 0.

(3) To reduce confusion, move the initialisation of len outside of: for
(len = 0; n > 0; n--) { since it doesn't have anything to do with the
loop counter n.

Change-Id: I6e381e753642e76118d16858d4095e6c3e378781
Reviewed-on: http://mcrd1-22-pc.corpnet.asus/code-review/master/256274
Reviewed-by: Carol_Jiang <carol_jiang@asus.com>
Tested-by: Carol_Jiang <carol_jiang@asus.com>