anthias: security: [2016-09-partner][CVE-2016-6689]
CVE-2016-6689 	A-30768347 	Moderate 	3.4, 3.10,
3.18 	Aug 9, 2016

An information disclosure vulnerability in Binder could enable a local
malicious application to access data outside of its permission levels.
This issue is rated as Moderate because it first requires compromising a
privileged process.

Additional technical details:

The interaction between the kernel /dev/binder and the usermode
Parcel.cpp means that when a Binder object is passed as
BINDER_TYPE_BINDER or BINDER_TYPE_WEAK_BINDER, a pointer to that object
(in the server process) is leaked to the client process as the cookie
value. This leads to a leak of a heap address in many of the privileged
Binder services, including system_server.

The fix is designed to zero out the Binder pointer and cookie before
sending it to the client process.

Change-Id: Ie68b7d84b987f0992bbf5cbff5a832c48f1d5ed0
Reviewed-on: http://mcrd1-22-pc.corpnet.asus/code-review/master/256386
Reviewed-by: Carol_Jiang <carol_jiang@asus.com>
Tested-by: Carol_Jiang <carol_jiang@asus.com>