common: security: prevent local malicious application to access data outside of its permission levels
The interaction between the kernel /dev/binder and the usermode Parcel.cpp
means that when a Binder object is passed as BINDER_TYPE_BINDER or
BINDER_TYPE_WEAK_BINDER, a pointer to that object (in the server process) is
leaked to the client process as the cookie value.
This leads to a leak of a heap address in many of the privileged
Binder services, including system_server.

CVE: CVE:2016-6689
Bug: 30768347
Change-Id: Id8baf79043bee64a55ca0342db8e1892c8ea2f7e
Reviewed-on: http://mcrd1-22-pc.corpnet.asus/code-review/master/257156
Reviewed-by: Yu-Hao Lin <yu-hao_lin@asus.com>
Tested-by: Yu-Hao Lin <yu-hao_lin@asus.com>
Reviewed-by: Wu, Meng-Hui <menghui_wu@asus.com>