security: binder: clear cookie
The interaction between the kernel /dev/binder and the usermode Parcel.cpp
means that when a Binder object is passed as BINDER_TYPE_BINDER or BINDER_TYPE_WEAK_BINDER,
a pointer to that object (in the server process) is leaked to the client process as the cookie value.
This leads to a leak of a heap address in many of the privileged Binder services, including system_server.

Change-Id: I43d8eef0b9adcb44d9f77ff9cfd730da14bbcee0