2016-09-partner: CVE-2016-3809: A-27532522: Information disclosure vulnerability in networking component
An information disclosure vulnerability in the networking component could enable a local
malicious application to access data outside of its permission levels.
This issue is rated as High because it could be used to access sensitive data without
explicit user permission.

The original fix was designed to not print kernel pointers to any user.
The updated fix is designed to not show empty tag stats for unprivileged uids as well.

Change-Id: I0191ae1543b650c1be0d526d2fcf0b690c065db8
Reviewed-on: http://mcrd1-22-pc.corpnet.asus/code-review/master/257856
Reviewed-by: joe_tsai <joe_tsai@asus.com>
Tested-by: joe_tsai <joe_tsai@asus.com>