smb-lib: support separate current settings from usb and pd
Add support to set current limits from usb and pd.

Ensure that the limit from the pd is only enforced when pd is active
and that the limit from usb is enforced only when pd is inactive.

There is one situation when the input needs to be suspended as
per the usb driver's request when pd sets system_suspend_supported.
For that allow the usb voter to set that low current value to
suspend.

Also ensure that the pd and usb votes are nullified when source
is removed.

In addition, the return the individual valid votes from usb and pd
when their current_max are read.

Bug: 67451182
Change-Id: I215e9e82f44061c4083a4df8cab400beadf6322d
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
Signed-off-by: Badhri Jagan Sridharan <Badhri@google.com>