sprat: binder: CVE-2016-6674
[CVE-2016-6674 ANDROID-30445380] Elevation of privilege vulnerability in system_server (device specific)
An elevation of privilege vulnerability in system_server could enable a local malicious application to execute arbitrary code within the context of a privileged process. This issue is rated as High because it could be used to gain local access to elevated capabilities, which are not normally accessible to a third-party application.

[Additional technical details]
The usage of weak references instead of strong references in Binder can potentially lead to a use-after-free vulnerability in system_server.
The fix is designed to no longer allow weak references in cases where strong references are needed.
Code snippet provided in bulletin patches as ANDROID-30445380

Signed-off-by: sanghyun.eom <sanghyun.eom@samsung.com>