drivers: touchscreen: CVE-2016-3940
Fix elevation of privilege vulnerability in Synaptics touchscreen
driver.
There is a possible stack overflow vulnerability in the
synaptics_rmi4_i2c_write function because the stack array size
is from user space.
The fix is designed to allocate heap memory for the temporary
buffer instead of stack memory to prevent the stack overflow
vulnerability.

Change-Id: I5a840ad9188bcfe1be601936924df53256ebecc3
Signed-off-by: Lianchao Song <songlianchao@huawei.com>