anthias: security: [2016-10-partner][CVE-2015-8963]
CVE-2015-8963	A-30952077
Upstream kernel	High	3.10, 3.18	Google internal
An elevation of privilege vulnerability in the kernel performance
subsystem could enable a local malicious application to execute
arbitrary code within the context of the kernel. This issue is rated as
High because it first requires compromising a privileged process.

Additional technical details:
A-30952077
There's a race condition on CPU unplug where the swevent hash array is
freed while it can still have events on. This will potentially lead to a
use-after-free vulnerability.

The fix is designed to not free the hash array on unplug and to have it
freed when the last swevent dies preventing the potential use-after-free
vulnerability.

Original patch link:
https://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=12ca6ad2e3a896256f086497a7c7406a547ee373

Original patch info.:
From: Peter Zijlstra <peterz@infradead.org>
Date: Tue, 15 Dec 2015 13:49:05 +0100
Subject: perf: Fix race in swevent hash

There's a race on CPU unplug where we free the swevent hash array
while it can still have events on. This will result in a
use-after-free which is BAD.

Simply do not free the hash array on unplug. This leaves the thing
around and no use-after-free takes place.

When the last swevent dies, we do a for_each_possible_cpu() iteration
anyway to clean these up, at which time we'll free it, so no leakage
will occur.

Change-Id: Ibe9f67b9bef39592cf80583ac2726936a6d65e67
Reviewed-on: http://mcrd1-22-pc.corpnet.asus/code-review/master/259900
Reviewed-by: allen1_huang <allen1_huang@asus.com>
Tested-by: allen1_huang <allen1_huang@asus.com>
Reviewed-by: Carol_Jiang <carol_jiang@asus.com>