perf: Fix race in swevent hash
There's a race on CPU unplug where we free the swevent hash array
while it can still have events on. This will result in a
use-after-free which is BAD.

Simply do not free the hash array on unplug. This leaves the thing
around and no use-after-free takes place.

When the last swevent dies, we do a for_each_possible_cpu() iteration
anyway to clean these up, at which time we'll free it, so no leakage
will occur.

CVE: CVE-2015-8963
Bug: 30952077

Change-Id: I1084e4e6d08e8cc9ed6a744091f2a6c1c6389201
Reviewed-on: http://mcrd1-22-pc.corpnet.asus/code-review/master/260348
Reviewed-by: Yu-Hao Lin <yu-hao_lin@asus.com>
Tested-by: Yu-Hao Lin <yu-hao_lin@asus.com>
Reviewed-by: Wu, Meng-Hui <menghui_wu@asus.com>