power: idtp9017: Do not cancel the tx online check when enabling TX
The online check keeps going to handle it properly when TX is enabling.
If the online check is doing in 3 seconds, actual check is deffered to
the next schduling time. If TX is enabled in case of this and the online
check is canceled, nobody will update the fake online.

Bug: 32397505
Change-Id: Id6fe4d6509e60819882359daf47abc5069a8ce53
Signed-off-by: kinam119.kim <kinam119.kim@lge.com>
Signed-off-by: Devin Kim <dojip.kim@lge.com>