swift: proc: prevent accessing /proc/<PID>/environ until it's ready
If /proc/<PID>/environ gets read before the envp[] array is fully set up
in create_{aout,elf,elf_fdpic,flat}_tables(), we might end up trying to
read more bytes than are actually written, as env_start will already be
set but env_end will still be zero, making the range calculation
underflow, allowing to read beyond the end of what has been written.

Fix this as it is done for /proc/<PID>/cmdline by testing env_end for
zero.  It is, apparently, intentionally set last in create_*_tables().

This bug was found by the PaX size_overflow plugin that detected the
arithmetic underflow of 'this_len = env_end - (env_start + src)' when
env_end is still zero.

The expected consequence is that userland trying to access
/proc/<PID>/environ of a not yet fully set up process may get
inconsistent data as we're in the middle of copying in the environment
variables.

CVE-2016-7916

Change-Id: I7b63c091e20e767965b7139e700c53da2fb5ea79
Reviewed-on: http://mcrd1-22-pc.corpnet.asus/code-review/master/261130
Reviewed-by: Jupiter Chen <jupiter_chen@asus.com>
Tested-by: Jupiter Chen <jupiter_chen@asus.com>
Reviewed-by: Wu, Meng-Hui <menghui_wu@asus.com>