common: ext4: fix potential use after free in __ext4_journal_stop
There is a potential use-after-free vulnerability in the __ext4_journal_stop function when the handle is freed in the first jbd2_journal_stop function because it references handle->h_err afterwards.
The fix is designed to store the handle->h_err value beforehand and avoid referencing potentially freed handle and preventing the potential use-after-free vulnerability.
CVE:CVE-2016-8961
Bug:30952474

Change-Id: I9abb9d0d49af4df450fd24e55a9979ad46dc3cf8
Reviewed-on: http://mcrd1-22-pc.corpnet.asus/code-review/master/261189
Reviewed-by: tc_tseng <tc_tseng@asus.com>
Tested-by: tc_tseng <tc_tseng@asus.com>
Reviewed-by: Wu, Meng-Hui <menghui_wu@asus.com>