power: idtp9017: Fix the race condition between TX enable and TX check
If the TX is enabled when fake online is true and device is off
from cradle, the TX check request is ignored. Thus the charging
state won't be updated until you put the device on the cradle or
the next online check is scheduled.

Do not ignore the TX check request to fix the race condition.
And rework the codes for the TX control.

Bug: 32478350
Change-Id: I7e1ac79b306d73904f4c32c1fb90c7d3da98c4d4
Signed-off-by: Devin Kim <dojip.kim@lge.com>
Signed-off-by: kinam119.kim <kinam119.kim@lge.com>