thermal: Export sensor batt_therm to upper layer
According to the hardware temperature test, the temperature of xo_therm 
grows quickly and can reach up to 50 or 60 degree in stability test.

The batt_therm grows slowly and  temperature growth trend is almost the 
same as skin temperature of the watch.

Change xo_therm to batt_therm to monitor the skin temperature.

Change-Id: I2a718d408fab0826f1317efd2f9c64bc4f07d3f3