CHG: Add fake charger function for factory test
In some test, the battery needs to stay normal when the enviroment 
temperature is high or the battery capacity is zero.

To meet the test requirement, the 'is_fake_battery' flag is added and 
can be modified by shell command. If the flag set, The CHG driver will 
check the flag and return normal values to upper layer.

The default value of the flag is false, and only used for test purpose.

Change-Id: Ide020e7786d258751d231f84a0373cefc7431871