USB: Enable charging anyway if SDP is detected.
Some wall chargers, in which USB D+/D- are not shortened,
could not be detected as floating chargers.
So, if SDP is detected but charging current is not configured
in 5 seconds, notify charger driver the 501mA charging current,
just 1mA more than standard 500mA value for Charger driver aware
of this issue.

Change-Id: I62ade2cb9be80a3b777af6354fa7935714d545e4