mpi: Fix NULL ptr dereference in mpi_powm() [ver #3]
If mpi_powm() is given a zero exponent, it wants to immediately return
either 1 or 0, depending on the modulus.  However, if the result was
initalised with zero limb space, no limbs space is allocated and a
NULL-pointer exception ensues.

Fix this by allocating a minimal amount of limb space for the result when
the 0-exponent case when the result is 1 and not touching the limb space
when the result is 0.

This affects the use of RSA keys and X.509 certificates that carry them.

Basically, this is a backport of a libgcrypt patch:
	http://git.gnupg.org/cgi-bin/gitweb.cgi?p=libgcrypt.git;a=patch;h=6e1adb05d290aeeb1c230c763970695f4a538526

Fixes: cdec9cb5167a ("crypto: GnuPG based MPI lib - source files (part 1)")
Signed-off-by: Andrey Ryabinin <aryabinin@virtuozzo.com>
Signed-off-by: David Howells <dhowells@redhat.com>
cc: Dmitry Kasatkin <dmitry.kasatkin@gmail.com>
cc: linux-ima-devel@lists.sourceforge.net
cc: stable@vger.kernel.org
Signed-off-by: James Morris <james.l.morris@oracle.com>

CVE: CVE-2016-8650
Bug: 33401771
Change-Id: Ia1f6ec102984547f3254c6087943ea5078fb7ddd
Reviewed-on: http://mcrd1-22-pc.corpnet.asus/code-review/master/286037
Reviewed-by: Yu-Hao Lin <yu-hao_lin@asus.com>
Tested-by: Yu-Hao Lin <yu-hao_lin@asus.com>
Reviewed-by: Wu, Meng-Hui <menghui_wu@asus.com>