arm/dt: afsf: Change gdsc clock to core_clk from core_root_clk
Changing of the gdsc node only impact on changing the source
from where we are deriving the clock. It is either root node
or the branch clock.
Deriving the clock from core_root_clock might be introducing delay
which can result in clock instability and causing the generation
of bad commands and interpretation by various GPU blocks.

Bug: 31495074
Change-Id: I115ab1d4b290d0a043f341a9079853b11685e4a1
Signed-off-by: neungq.park <neungq.park@lge.com>
Signed-off-by: Devin Kim <dojip.kim@lge.com>