wlan: Replace time of the day with jiffies during scan reject
Scan reject changes wrongly used time of the day timestamp for
last_scan_reject_timestamp which leads to device crash
if time is changed for device in between.

Add changes to use jiffies instead of time of the day and
intialize scan_reject_reason and scan_reject_session_id with proper values.

Change-Id: I113042726569525c5b39c1861e9004e0217d34b2
CRs-Fixed: 1099854