prima: Fix array out-of-bounds & integer underflow in _iw_set_genie
qcacld-2.0 to prima propagation

'wrqu->data.length' holds the total number of IE data buffer.
Add a check to make sure the number of remaining data to be read is
greater than or equal to IE length.
Also, advance the buffer pointer to point to the next element only
if next element is present.

Change-Id: Ia98474749a9e3f3cb62e2f988de5409f1c6ddcee
CRs-Fixed: 1100132