power: charger: fix a bug about charger type
The msm_otg_probe sets charger type to USB and healthd will detect the charger type.
If the type is unknown, healthd indicates that there is no charger driver registered
and display charger parameters wrongly, for example, chg type is empty.
The issue happens as follows:
1), when usb is connected, the watch is booting and the type is USB.
2), before the starting of healthd, disconnect the usb and the type is unknown.
So set charger type to USB when usb is disconnected.

Change-Id: Ia96b257cd958aea224159c9ce08fb4081981b44b