BACKPORT: crypto: arm/aes - replace scalar AES cipher
This replaces the scalar AES cipher that originates in the OpenSSL project
with a new implementation that is ~15% (*) faster (on modern cores), and
reuses the lookup tables and the key schedule generation routines from the
generic C implementation (which is usually compiled in anyway due to
networking and other subsystems depending on it).

Note that the bit sliced NEON code for AES still depends on the scalar cipher
that this patch replaces, so it is not removed entirely yet.

* On Cortex-A57, the performance increases from 17.0 to 14.9 cycles per byte
  for 128-bit keys.

Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>

(cherry picked from commit 81edb42629758bacdf813dd5e4542ae26e3ad73a)

Conflicts:
	arch/arm/crypto/Kconfig

Bug: 112008522
Test: As series, see Ic61c13b53facfd2173065be715a7ee5f3af8760b
Change-Id: Iac145d586b01c5c18f3cb5979bf6035232d4dbf2
Signed-off-by: Eric Biggers <ebiggers@google.com>
(cherry picked from commit b7912e0f1d862b8a7dce4f0f1d27411181ed3bce)
Signed-off-by: Chandrasekhar Mattaparthy <cmatta@codeaurora.org>