ALSA: info: Check for integer overflow in snd_info_entry_write()
CVE-2017-0404
snd_info_entry_write() resizes the buffer with an unsigned long
size argument that gets truncated because resize_info_buffer()
takes the size parameter as an unsigned int. On 64-bit kernels,
this causes the following copy_to_user() to write out-of-bounds
if (pos + count) can't be represented by an unsigned int.

Bug: 32510733
Change-Id: Ib7f341fbdeae2c4a5676beda55357d0fc985b97f