ANDROID: sdcardfs: add support for user permission isolation
This allows you to hide the existence of a package from
a user by adding them to an exclude list. If a user
creates that package's folder and is on the exclude list,
they will not see that package's id.

Bug: 34542611
Change-Id: I9eb82e0bf2457d7eb81ee56153b9c7d2f6646323
Signed-off-by: Daniel Rosenberg <drosen@google.com>
(cherry picked from commit 4c3307fea0e8da10723325ee2492bba3c43c02d3)
Signed-off-by: Manjunatha Madana <mamanj@codeaurora.org>