qcacld-2.0: Fix flooding of DPTRACE logs for DHCP/EAPOL pkts
For DHCP and EAPOL packets, the APIs to find whether it is a DHCP
or EAPOL packet returns A_STATUS_OK and A_STATUS_FAILED instead of
boolean true and false. This leads to wrong identifiaction of packets
in rx path as DHCP and EAPOL packets and so each rx packet  is either
identified as DHCP or EAPOL packet and hence flooding of logs in kmsg.

Fix the return type of APIs to return boolean true and false.

Change-Id: Ic8ca56690a17ef1b1196a243d75d2edf5f735a09
CRs-Fixed: 1062897
Bug: 34691311
Signed-off-by: Srinivas Girigowda <sgirigow@codeaurora.org>