input: synaptics_dsx: Refactor HTC synaptics driver
This patch is to enable HTC synaptics driver to be able to be built into
kernel module. This removes an issue with duplicated trace define shared
with the clock subsystem as well as seperating the object naming so that
both the HTC and the stock versions can be picked up.

Bug: 32574003

Signed-off-by: David Lin <dtwlin@google.com>
Change-Id: Ie949b4fca0e490b0ed151ba6eedb4efee7c6f33f