ext4 crypto: remove incorrect __exit attributes
Cleans up a warning from modpost sugguesting that these be removed.

WARNING: vmlinux.o(.init.text+0x2a330): Section mismatch in reference
from the function pfk_init() to the function
.exit.text:pfk_ecryptfs_deinit()
The function __init pfk_init() references
a function __exit pfk_ecryptfs_deinit().
This is often seen when error handling in the init function
uses functionality in the exit path.
The fix is often to remove the __exit annotation of
pfk_ecryptfs_deinit() so it may be used outside an exit section.

Bug: 34712722
Change-Id: I02abdb1d225bf6c681c1cdb4f4edd0cc3f8c0cfd