misc: mnh: fix resume sequence
Certain config registers are written by the platform driver after
enumeration. However, after resume, these values are reset and the
device is not enumerated again. Call the pci framework APIs to save and
restore endpoint state after resume.

Also, remove ready signal. This gets asserted after we release the reset
which occurs in the pci platform driver.

Also, correct the order in which I enable and disable the power supplies
to match the spec and reference board.

Change-Id: Iffaf7189963b268541ce91c1bb0130e1633cc969
Signed-off-by: Trevor Bunker <trevorbunker@google.com>