wlan: Fix buffer overflow in WLANSAP_Set_WPARSNIes()
qcacld-2.0 to prima propagation

Currently In WLANSAP_Set_WPARSNIes() the parameter WPARSNIEsLen
is user-controllable and never validates which uses as the length
for a memory copy. This enables user-space applications to corrupt
heap memory and potentially crash the kernel.

Fix is to validate the WPARSNIes length to its max before use as the
length for a memory copy.

Change-Id: Ifa8413c8f5729cb61728e579500e9934cd476ffb
CRs-Fixed: 1102648