Reverse cpu offline sequence
In the original design, cpu0 and cpu1 will set to offline
when system is not busy.
But CTS item record_cmd#cpu_option and stat_cmd#cpu_option
in CtsSimpleperfTestCase will fail when cpu0 is not online.
So we reverse the sequence when system want to set cpu
offline to let cpu2 and cpu3 get offline first.

Change-Id: I2eb8497b81d98219bfd4a65350ee0989dd4ae272