power: battery: move batfet reset into a work queue task
Resetting the batfet takes a nontrivial amount of time (hundreds of ms).
When the batfet is reset on the same thread responsible for setting the
display brightness, there is a noticeable pause in the otherwise smooth
brightness transition. Move the operation to a work queue item so it does
not affect the screen brightness.

Bug: 34083323
Change-Id: Ic5281e5ec4aae16606105d7d7bc060e498f91f34
Signed-off-by: Steve Pfetsch <spfetsch@google.com>